+++ test.py
@@ -6,13 +6,7 @@ from scipy import interpolate
 from netCDF4 import Dataset
 from multiprocessing import Pool
 import time
-
-def correlation_coef(x,y):
-    
-    r = (np.sum(((x-np.mean(x))*(y-np.mean(y)))))/(np.sqrt(np.sum(np.square(x-np.mean(x)))*np.sum(np.square(y-np.mean(y)))))
-
-    return r
-
+import pandas as pd
 
 def Ux_it_offset(it):
 
@@ -76,70 +70,17 @@ with Pool() as pool:
         print(it,time.time()-start_time)
         it+=1
     Ux = np.array(Ux)
+    
 
-import csv 
-    
-# field names 
-fields = ['velocityx'] 
-    
-# data rows of csv file 
-rows = [Ux] 
-    
 # name of csv file 
 filename = in_dir+"velocityx.csv"
     
-# writing to csv file 
-with open(filename, 'w') as csvfile: 
-    # creating a csv writer object 
-    csvwriter = csv.writer(csvfile) 
-        
-    # writing the fields 
-    csvwriter.writerow(fields) 
-        
-    # writing the data rows 
-    csvwriter.writerows(rows)
+dq = dict()
+
+dq["velocityx"] = Ux
+
+df = pd.DataFrame(dq)
+
+df.to_csv(filename)
 
 print("line 102",time.time()-start_time)
-
-a = Dataset(in_dir+"Dataset.nc")
-
-Time_OF = np.array(a.variables["time_OF"])
-Time_sampling = np.array(a.variables["time_sampling"])
-Time_sampling = Time_sampling - Time_sampling[0]
-
-Time_start = 0
-Time_end = Time_sampling[-1]
-
-dt = Time_OF[1] - Time_OF[0]
-
-Time_start_idx = np.searchsorted(Time_OF,Time_start)
-Time_end_idx = np.searchsorted(Time_OF,Time_end)
-
-Time_OF = Time_OF[Time_start_idx:Time_end_idx]
-
-RtAeroMxh = np.array(a.variables["RtAeroMxh"][Time_start_idx:Time_end_idx])
-
-print("line 121",time.time()-start_time)
-
-
-f = interpolate.interp1d(Time_sampling,velocityx)
-Ux = f(Time_OF)
-
-fig,ax = plt.subplots(figsize=(14,8))
-
-
-corr = correlation_coef(RtAeroMxh,Ux)
-corr = round(corr,2)
-
-ax.plot(Time_OF,RtAeroMxh,"-b")
-ax.set_ylabel("Mx",fontsize=14)
-
-ax2=ax.twinx()
-ax2.plot(Time_OF,Ux,"-r")
-ax2.set_ylabel("x velocity",fontsize=14)
-
-plt.title("Correlation = {0}".format(corr),fontsize=16)
-
-ax.set_xlabel("Time [s]",fontsize=16)
-plt.tight_layout()
-plt.show()
